Game is finished, came up with the idea to split the server into a lobby
where people chill and a game server with the game itself to play on.
This means I have to divide the protocol between the servers, things you
can do in the lobby and things you can do in the game (for instance,
CHAT is Lobby/Game and Move is only Game).